Scripts_Data
This folder contain an example of the script for calculating GL and LL values from an example of data provided. 

Note: the data provided is a random example of numbers and a random creation of polygons. It does not represent the reality of the study case. Study case data cannot be shared due to data sharing agreement policy with data providers.